+++ app/scripts/data-fetchers/genbank-fetcher.js
@@ -124,10 +124,9 @@ class GBKDataFetcher {
     this.errorTxt = '';
 
     this.dataPromise = fetch(dataConfig.url, {
-      headers: {
-        'Content-Encoding': gzipped ? 'gzip' : 'text/plain'
-      },
-      mode: 'cors'
+      mode: 'cors',
+      redirect: 'follow',
+      method: 'GET'
     })
       .then(response => (gzipped ? response.arrayBuffer() : response.text()))
       .then(buffer => {